+++ dotfiles/.config/hypr/scripts/keybindings.sh
@@ -37,4 +37,4 @@ BEGIN {
 
     # Output: Line 1 (Keys), Line 2 (Description), followed by the Null separator
     printf "%s\n➔ %s\0", combo, desc
-}' | rofi -dmenu -replace -p "Keybinds" -sep '\0' -eh 2 -config ~/.config/rofi/config-compact.rasi
+}' | rofi -dmenu -i -replace -p "Keybinds" -sep '\0' -eh 2 -config ~/.config/rofi/config-compact.rasi