+++ CHANGELOG.md
@@ -2,7 +2,7 @@ Version 2.14.1
 --------------------------------------------------------
 - You can keep the status bar expanded with new setting "alwaysExpanded": true for the bar section
 - Flexible width for the statusbar if the width of "expandedWidth": 680 is not enough.
-- New ststus bar module for volume control
+- New status bar module for volume control
 - New keybinding to reload the statusbar SUPER + ALT + B
 
 Version 2.14.0