+++ CHANGELOG.md
@@ -3,8 +3,8 @@ Version 2.14.1
 - The Quickshell statusbar can be configured in ~/.config/ml4w-statusbar/statusbar.json The file must be created individually. The full list of configuration options are available here: https://ml4w.com/os/ml4w-apps/statusbar
 - You can keep the status bar expanded with new setting "alwaysExpanded": true for the bar section. Also available as a toggle switch in the sidebar.
 - Flexible width for the statusbar if the width of "expandedWidth": 680 is not enough.
-- New status bar module for volume control
-- New keybinding to reload the statusbar SUPER + ALT + B
+- New status bar modules: Volume control, battery status, power profile plus submenu
+- Status Bar (Waybar or Quickshell) can be reloaded with SUPER + SHIFT + B and toggled with SUPER + CTRL + B
 - Added Network Manager Applet (nm-applet) to ml4w-autostart
 - Rofi menues can be closed by clicking outside of Rofi.